Skiller logikk ut fra FullførKjørelistebehandlingSteg til FullførKjørelistebehandlingService, slik at sistnevnte brukes både fra automatisk og manuell behandling av kjøreliste
